+++ pcs/common/pacemaker/constraint/__init__.py
@@ -1,7 +1,6 @@
 from .all import (
     CibConstraintsDto,
     get_all_constraints_ids,
-    get_all_location_constraints_ids,
     get_all_location_rules_ids,
 )
 from .colocation import (